pyproject.toml
[project]
name = "aiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.46.0",
    "lorem>=0.1.1",
    "python-dotenv>=1.1.1",
]
